Falta:(miscartas/miseventos)/notif

+++ app/src/main/java/com/example/magickfinaljesus/Cartas.kt
@@ -5,7 +5,7 @@ import java.io.Serializable
 data class Cartas(
     val id:String?=null,
     val nombre:String?=null,
-    val precio:Int?=null,
+    val precio:Float?=null,
     val disponible:Boolean?=null,
     val img:String?=null,
     val color:String?=null):Serializable
